[hotfix] #35 ProjectUpdate 관련 DTO 리팩토링
